src/requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11
django-registration-redux==2.0
django-rest-framework==0.1.0
djangorestframework==3.7.3
idna==2.6
oauthlib==2.0.6
PyJWT==1.5.3
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
urllib3==1.22
